requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
async-timeout==4.0.3
asyncpg==0.30.0
attrs==25.4.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
comm==0.2.3
coverage==7.11.0
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.3.0
executing==2.2.1
filetype==1.2.0
frozenlist==1.8.0
gemini_ai==0.0.5
google-ai-generativelanguage==0.9.0
google-api-core==2.28.1
google-api-python-client==2.186.0
google-auth==2.42.1
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.71.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
iniconfig==2.3.0
ipython==8.37.0
ipywidgets==8.1.8
jedi==0.19.2
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
jupyterlab_widgets==3.0.16
langchain==1.0.3
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.2
langchain-google-genai==3.0.0
langchain-openai==1.0.1
langchain-postgres==0.0.16
langchain-text-splitters==1.0.0
langgraph==1.0.2
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.39
marshmallow==3.26.1
matplotlib-inline==0.2.1
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.2.6
openai==2.6.1
orjson==3.11.4
ormsgpack==1.11.0
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pgvector==0.3.6
pillow==12.0.0
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psycopg==3.2.12
psycopg-binary==3.2.12
psycopg-pool==3.2.7
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.2.5
pypdf==6.1.3
pytest==8.4.2
pytest-cov==7.0.0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.44
stack-data==0.6.3
tenacity==9.1.2
tiktoken==0.12.0
tomli==2.3.0
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.14
widgetsnbextension==4.0.15
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
